pyproject.toml
[tool.poetry]
name = "compiler-workflows"
version = "0.0.1"
description = "Compiler Explorer workflows and UI"
authors = ["Matt Godbolt <matt@godbolt.org>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.3"
boto3 = "^1.20.5"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.10b0"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
